requirements.txt
numpy
pandas
matplotlib
opencv-python
scipy
Pillow
gradio

